refactor: split performance utils
