+++ packages/utils/performance/debounce.ts
@@ -1,0 +1,13 @@
+/**
+ * Debounce function calls to reduce unnecessary executions
+ */
+export function debounce<T extends (...args: unknown[]) => void>(
+  func: T,
+  delay: number
+): (...args: Parameters<T>) => void {
+  let timeoutId: ReturnType<typeof setTimeout>;
+  return function (this: unknown, ...args: Parameters<T>) {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => func.apply(this as never, args), delay);
+  };
+}
